auth feature complete and some additional files were created and fixed pk parameters to long type and .env has credentials of system admin.

+++ src/main/java/com/BezuhlyiBohdanK22_1/qualitrack/entity/AcademicRankEntity.java
@@ -19,7 +19,7 @@ public class AcademicRankEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "lecture_academic_rank_id")
-    private Integer lectureAcademicRankId;
+    private Long lectureAcademicRankId;
 
     @NotNull
     @ManyToOne(fetch = FetchType.LAZY)
